improve process handling
