+++ Cargo.toml
@@ -68,3 +68,4 @@ mockall = "0.14"
 mockito = "1.7"
 serial_test = "3"
 tempfile = "3.25"
+sysinfo = "0.38"
